Update cheapino firmware
* Extract keymap definitions to follow the external userspace model. A
  default keymap should probably be added again as an example.
* Move configuration to keyboard.json.
* Enable LTO.
* Move encoder button handling to the keymap for full qmk feature
  support (layers, mod-tap).
* Inject encoder turn events into the qmk encoder pipeline, with the
  same motivation as above.
* Rename files to avoid clashing with qmk-internal files (encoder.h).
* Faster matrix store/compare primitives.
