--- keyboards/cheapino/encoder.h
@@ -1,5 +1,0 @@
-//
-// Created by Thomas Haukland on 25/03/2023.
-//
-
-void fix_encoder_action(matrix_row_t current_matrix[]);
