--- keyboards/cheapino/halconf.h
@@ -1,8 +1,0 @@
-
-#pragma once
-
-#define HAL_USE_PWM TRUE
-#define HAL_USE_PAL TRUE
-#define HAL_USE_I2C TRUE
-
-#include_next <halconf.h>
