+++ keyboards/cheapino/matrix-encoder.h
@@ -1,0 +1,6 @@
+//
+// Created by Thomas Haukland on 25/03/2023.
+//
+#include "quantum.h"
+
+void fix_encoder_action(matrix_row_t current_matrix[]);
