+++ keyboards/cheapino/rules.mk
@@ -1,8 +1,3 @@
-CAPS_WORD_ENABLE = yes
-CUSTOM_MATRIX = lite
-WS2812_DRIVER = vendor
-RGBLIGHT_ENABLE = yes
-DEFERRED_EXEC_ENABLE = yes
-SRC += encoder.c
-SRC += ghosting.c
+SRC += matrix-encoder.c
+SRC += matrix-ghosting.c
 SRC += matrix.c
